+++ webworker.js
@@ -4,6 +4,19 @@ const stdout = []
 const stderr = []
 let exitCode = 0
 
+
+function ChangeUrl(url, fileName)
+{
+    let splitUrl = url.split('/')
+    splitUrl[splitUrl.length -1] = fileName
+    return splitUrl.join('/')
+}
+// const magickJsCurrentPath = 'https://knicknic.github.io/wasm-imagemagick/magick.js';
+function GetCurrentUrlDifferentFilename(fileName)
+{
+    return ChangeUrl(magickJsCurrentPath, fileName)
+}
+
 if (typeof Module == 'undefined') {
   Module = {
     noInitialRun: true,
@@ -21,6 +34,9 @@ if (typeof Module == 'undefined') {
     quit: status=> {
       exitCode = status
     }
+  }
+  if(typeof magickJsCurrentPath !== "undefined"){
+      Module.locateFile = GetCurrentUrlDifferentFilename;
   }
 
   // see https://kripken.github.io/emscripten-site/docs/api_reference/module.html